[rt_draw.c]: Fixing wrong cast, affecting the Apogee logo animation. (#31)
* [rt_draw.c]: Fixing wrong cast, affecting the Apogee logo animation.

* cast scale factors back to int

---------

+++ rott/rt_draw.c
@@ -3071,8 +3071,8 @@ void ApogeeTitle (void)
    y=APOGEESTARTY<<16;
    dy=((APOGEEENDY-APOGEESTARTY)<<16)/time;
 
-   dscale=((APOGEESCALEEND-APOGEESCALESTART)<<16)/time;
-   scale=APOGEESCALESTART<<16;
+   dscale=(int)((APOGEESCALEEND-APOGEESCALESTART)<<16)/time;
+   scale=(int)(APOGEESCALESTART<<16);
 
    angle=0;
    dangle=(FINEANGLES<<17)/time;
